# swiss: add correctness tests and improve benchmarking