```
name                                         old time/op  new time/op  delta
StringMap/avgLoad,n=10/Map/Get-10            9.53ns ± 6%  8.43ns ± 1%  -11.50%  (p=0.000 n=10+9)
StringMap/maxLoad,n=13/Map/Get-10            10.0ns ± 2%   9.2ns ±26%   -8.35%  (p=0.008 n=8+10)
StringMap/minLoad,n=14/Map/Get-10            9.10ns ± 4%  8.40ns ± 0%   -7.65%  (p=0.000 n=9+8)
StringMap/avgLoad,n=83/Map/Get-10            11.0ns ± 9%   9.2ns ±11%  -16.57%  (p=0.000 n=10+10)
StringMap/maxLoad,n=111/Map/Get-10           11.1ns ± 7%   9.8ns ±10%  -11.34%  (p=0.000 n=10+10)
StringMap/minLoad,n=112/Map/Get-10           11.5ns ± 9%   8.5ns ± 4%  -25.56%  (p=0.000 n=10+10)
StringMap/avgLoad,n=671/Map/Get-10           15.7ns ± 3%   9.0ns ± 3%  -42.31%  (p=0.000 n=10+10)
StringMap/maxLoad,n=895/Map/Get-10           16.8ns ± 2%  10.3ns ± 3%  -38.98%  (p=0.000 n=10+10)
StringMap/minLoad,n=896/Map/Get-10           16.7ns ± 5%   8.6ns ± 2%  -48.90%  (p=0.000 n=10+10)
StringMap/avgLoad,n=5375/Map/Get-10          25.8ns ± 1%   9.3ns ± 1%  -63.88%  (p=0.000 n=10+10)
StringMap/maxLoad,n=7167/Map/Get-10          25.1ns ± 1%  11.2ns ± 1%  -55.28%  (p=0.000 n=9+8)
StringMap/minLoad,n=7168/Map/Get-10          25.2ns ± 1%   8.9ns ± 1%  -64.77%  (p=0.000 n=10+10)
StringMap/avgLoad,n=86015/Map/Get-10         30.5ns ± 1%  10.9ns ± 2%  -64.34%  (p=0.000 n=9+10)
StringMap/maxLoad,n=114687/Map/Get-10        32.4ns ± 1%  13.7ns ± 1%  -57.65%  (p=0.000 n=10+9)
StringMap/minLoad,n=114688/Map/Get-10        32.5ns ± 1%  14.7ns ± 2%  -54.82%  (p=0.000 n=10+10)
Int64Map/avgLoad,n=10/Map/Get-10             5.11ns ± 3%  4.85ns ± 1%   -5.13%  (p=0.000 n=10+10)
Int64Map/maxLoad,n=13/Map/Get-10             5.39ns ± 6%  4.94ns ± 2%   -8.34%  (p=0.000 n=10+10)
Int64Map/minLoad,n=14/Map/Get-10             4.86ns ± 3%  4.83ns ± 0%     ~     (p=0.796 n=10+10)
Int64Map/avgLoad,n=83/Map/Get-10             5.23ns ± 3%  5.18ns ± 7%     ~     (p=0.529 n=10+10)
Int64Map/maxLoad,n=111/Map/Get-10            4.94ns ± 1%  5.38ns ±10%   +8.85%  (p=0.000 n=10+10)
Int64Map/minLoad,n=112/Map/Get-10            5.10ns ± 5%  4.99ns ± 5%     ~     (p=0.247 n=10+10)
Int64Map/avgLoad,n=671/Map/Get-10            6.03ns ± 7%  5.36ns ± 5%  -11.08%  (p=0.000 n=10+10)
Int64Map/maxLoad,n=895/Map/Get-10            6.42ns ± 8%  5.62ns ± 3%  -12.48%  (p=0.000 n=10+10)
Int64Map/minLoad,n=896/Map/Get-10            6.21ns ± 5%  5.06ns ± 6%  -18.53%  (p=0.000 n=10+10)
Int64Map/avgLoad,n=5375/Map/Get-10           18.3ns ± 2%   5.7ns ± 2%  -68.76%  (p=0.000 n=10+10)
Int64Map/maxLoad,n=7167/Map/Get-10           18.9ns ± 1%   6.4ns ± 3%  -66.11%  (p=0.000 n=10+10)
Int64Map/minLoad,n=7168/Map/Get-10           18.8ns ± 2%   5.4ns ± 1%  -71.16%  (p=0.000 n=10+10)
Int64Map/avgLoad,n=86015/Map/Get-10          23.9ns ± 1%   6.9ns ± 0%  -71.24%  (p=0.000 n=10+9)
Int64Map/maxLoad,n=114687/Map/Get-10         22.6ns ± 1%   9.4ns ± 1%  -58.58%  (p=0.000 n=9+9)
Int64Map/minLoad,n=114688/Map/Get-10         22.6ns ± 1%   6.7ns ± 1%  -70.19%  (p=0.000 n=10+10)

name                                         old time/op  new time/op  delta
StringMap/avgLoad,n=10/Map/PutDelete-10      26.3ns ±11%  23.3ns ± 2%  -11.41%  (p=0.000 n=10+8)
StringMap/maxLoad,n=13/Map/PutDelete-10      27.4ns ± 9%  23.8ns ± 4%  -12.84%  (p=0.000 n=10+9)
StringMap/minLoad,n=14/Map/PutDelete-10      25.0ns ±10%  23.5ns ± 3%   -5.98%  (p=0.001 n=10+9)
StringMap/avgLoad,n=83/Map/PutDelete-10      31.6ns ± 7%  23.4ns ± 4%  -25.94%  (p=0.000 n=10+10)
StringMap/maxLoad,n=111/Map/PutDelete-10     31.2ns ± 4%  23.6ns ± 3%  -24.30%  (p=0.000 n=10+10)
StringMap/minLoad,n=112/Map/PutDelete-10     31.0ns ± 3%  23.6ns ± 4%  -23.60%  (p=0.000 n=10+10)
StringMap/avgLoad,n=671/Map/PutDelete-10     45.2ns ± 1%  23.5ns ± 1%  -47.96%  (p=0.000 n=10+9)
StringMap/maxLoad,n=895/Map/PutDelete-10     44.8ns ± 2%  23.2ns ± 1%  -48.23%  (p=0.000 n=10+7)
StringMap/minLoad,n=896/Map/PutDelete-10     44.6ns ± 2%  23.6ns ± 2%  -47.11%  (p=0.000 n=10+8)
StringMap/avgLoad,n=5375/Map/PutDelete-10    56.7ns ± 1%  24.3ns ± 3%  -57.25%  (p=0.000 n=10+10)
StringMap/maxLoad,n=7167/Map/PutDelete-10    53.1ns ± 1%  25.1ns ± 5%  -52.71%  (p=0.000 n=10+10)
StringMap/minLoad,n=7168/Map/PutDelete-10    53.2ns ± 1%  24.9ns ± 1%  -53.22%  (p=0.000 n=10+8)
StringMap/avgLoad,n=86015/Map/PutDelete-10   60.9ns ± 0%  38.9ns ± 3%  -36.17%  (p=0.000 n=9+10)
StringMap/maxLoad,n=114687/Map/PutDelete-10  62.4ns ± 2%  42.2ns ± 0%  -32.32%  (p=0.000 n=10+7)
StringMap/minLoad,n=114688/Map/PutDelete-10  62.7ns ± 1%  42.1ns ± 1%  -32.82%  (p=0.000 n=10+10)
Int64Map/avgLoad,n=10/Map/PutDelete-10       18.4ns ± 9%  15.8ns ±12%  -13.99%  (p=0.000 n=10+10)
Int64Map/maxLoad,n=13/Map/PutDelete-10       19.4ns ± 2%  15.6ns ± 7%  -19.61%  (p=0.000 n=8+9)
Int64Map/minLoad,n=14/Map/PutDelete-10       16.8ns ± 9%  15.0ns ± 4%  -10.85%  (p=0.000 n=9+10)
Int64Map/avgLoad,n=83/Map/PutDelete-10       19.6ns ± 4%  14.7ns ± 1%  -25.14%  (p=0.000 n=9+8)
Int64Map/maxLoad,n=111/Map/PutDelete-10      18.7ns ± 7%  14.5ns ± 2%  -22.64%  (p=0.000 n=10+10)
Int64Map/minLoad,n=112/Map/PutDelete-10      18.4ns ± 4%  14.4ns ± 3%  -21.65%  (p=0.000 n=10+10)
Int64Map/avgLoad,n=671/Map/PutDelete-10      27.1ns ± 2%  14.2ns ± 3%  -47.52%  (p=0.000 n=10+9)
Int64Map/maxLoad,n=895/Map/PutDelete-10      27.7ns ± 1%  14.3ns ± 2%  -48.26%  (p=0.000 n=9+10)
Int64Map/minLoad,n=896/Map/PutDelete-10      27.7ns ± 2%  14.2ns ± 2%  -48.62%  (p=0.000 n=10+10)
Int64Map/avgLoad,n=5375/Map/PutDelete-10     44.4ns ± 1%  16.0ns ± 2%  -63.93%  (p=0.000 n=10+8)
Int64Map/maxLoad,n=7167/Map/PutDelete-10     43.0ns ± 1%  17.1ns ± 5%  -60.31%  (p=0.000 n=10+10)
Int64Map/minLoad,n=7168/Map/PutDelete-10     43.0ns ± 1%  16.9ns ± 3%  -60.68%  (p=0.000 n=10+10)
Int64Map/avgLoad,n=86015/Map/PutDelete-10    50.6ns ± 0%  21.6ns ± 3%  -57.41%  (p=0.000 n=9+10)
Int64Map/maxLoad,n=114687/Map/PutDelete-10   47.3ns ± 0%  21.8ns ± 2%  -53.89%  (p=0.000 n=8+9)
Int64Map/minLoad,n=114688/Map/PutDelete-10   47.4ns ± 1%  22.1ns ± 3%  -53.39%  (p=0.000 n=10+10)
```

<!-- Reviewable:start -->
- - -
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/cockroachdb/swiss/3)
<!-- Reviewable:end -->